+++ src/stack/repos/list.py
@@ -1,4 +1,3 @@
-# Copyright © 2022, 2023 Vulcanize
 # Copyright © 2025 Bozeman Pass, Inc.
 
 # This program is free software: you can redistribute it and/or modify
@@ -14,67 +13,16 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http:#www.gnu.org/licenses/>.
 
-# env vars:
-# STACK_REPO_BASE_DIR defaults to ~/bpi
-
-
 import click
 
-from stack.config.util import get_dev_root_path
-from stack.deploy.stack import Stack
-from stack.opts import opts
-from stack.util import error_exit
+from stack.repos.list_stack import command as list_stack
 
 
-def locate_stacks_beneath(search_path=get_dev_root_path()):
-    stacks = []
-    if search_path.exists():
-        for path in search_path.rglob("stack.yml"):
-            stacks.append(Stack().init_from_file(path))
-
-    return stacks
-
-
-def locate_single_stack(stack_name, search_path=get_dev_root_path()):
-    stacks = locate_stacks_beneath(search_path)
-    candidates = [s for s in stacks if s.name == stack_name]
-    if len(candidates) == 1:
-        return candidates[0]
-    elif len(candidates) > 1:
-        error_exit(f"multiple stacks named {stack_name} found")
-    return None
-
-
-@click.command()
-@click.argument("stack-locator", required=False)
-@click.option("--names-only", is_flag=True, default=False, help="only show stack names")
-@click.option("--paths-only", is_flag=True, default=False, help="only show stack paths")
+@click.group()
 @click.pass_context
-def command(ctx, stack_locator, names_only, paths_only):
-    """list stacks in a repository"""
-    dev_root_path = get_dev_root_path()
-    if opts.o.verbose:
-        print(f"Dev Root is: {dev_root_path}")
+def command(ctx):
+    """list available stacks, etc."""
+    pass
 
-    search_path = dev_root_path
 
-    stacks = locate_stacks_beneath(search_path)
-    stacks.sort(key=lambda s: s.name)
-
-    if stack_locator:
-        stacks = [s for s in stacks if stack_locator in str(s.file_path) or stack_locator in s.name]
-
-    max_name_len = 0
-    max_path_len = 0
-    for stack in stacks:
-        max_name_len = max(max_name_len, len(stack.name))
-        max_path_len = max(max_path_len, len(str(stack.file_path.parent)))
-
-    padding = 8
-    for stack in stacks:
-        if names_only:
-            print(stack.name)
-        elif paths_only:
-            print(str(stack.file_path.parent))
-        else:
-            print(f"{stack.name.ljust(max_name_len + padding)} {str(stack.file_path.parent).ljust(max_path_len + 8)}")
+command.add_command(list_stack, "stacks")
