+++ src/stack/repos/setup_repositories.py
@@ -31,9 +31,9 @@ from stack.build.build_util import get_containers_in_scope, host_and_path_for_re
 from stack.config.util import get_config_setting, get_dev_root_path
 from stack.deploy.stack import get_parsed_stack_config
 from stack.opts import opts
+from stack.repos.list_stack import resolve_stack
 from stack.util import (
     is_git_repo,
-    check_if_stack_exists,
     include_exclude_check,
     error_exit,
 )
@@ -193,10 +193,9 @@ def command(ctx, stack, include, exclude, git_ssh, check_only, pull, branches):
 
     if not stack:
         stack = ctx.obj.stack_path
-    check_if_stack_exists(stack)
 
+    stack_config = resolve_stack(stack)
     required_stacks = []
-    stack_config = get_parsed_stack_config(stack)
     if stack_config.is_super_stack():
         for stack_refs in stack_config.get_required_stacks():
             try:
@@ -205,7 +204,7 @@ def command(ctx, stack, include, exclude, git_ssh, check_only, pull, branches):
                 error_exit(f"\n******* git command returned error exit status:\n{error}")
             required_stacks.append(os.path.sep.join([repo_path, stack_refs[constants.path_key]]))
     else:
-        required_stacks.append(stack)
+        required_stacks.append(stack_config)
 
     for stack in required_stacks:
         branches_array = []
@@ -218,7 +217,7 @@ def command(ctx, stack, include, exclude, git_ssh, check_only, pull, branches):
 
         dev_root_path = get_dev_root_path()
 
-        if not quiet:
+        if verbose:
             print(f"Dev Root is: {dev_root_path}")
 
         if not os.path.isdir(dev_root_path):
@@ -241,7 +240,8 @@ def command(ctx, stack, include, exclude, git_ssh, check_only, pull, branches):
                 print(f"Stack: {stack}")
 
         if not repos_in_scope:
-            print(f"WARN: stack {stack} does not define any repositories")
+            if verbose:
+                print(f"NOTE: stack {stack} does not define any repositories")
             continue
 
         repos = []
